third_party/WebKit/LayoutTests/external/wpt/html/browsers/the-window-object/window-indexed-properties-expected.txt

This is a testharness.js-based test.
PASS Indexed properties of the window object (non-strict mode)
PASS Ensure indexed properties have the correct configuration
FAIL Indexed properties of the window object (non-strict mode) 1 assert_equals: expected false but got true
PASS Indexed properties of the window object (non-strict mode) 2
PASS Indexed properties of the window object (non-strict mode) 3
Harness: the test ran to completion.
